requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
Django==4.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-rest-passwordreset==1.5.0
django-storages==1.14.4
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2